Add files via upload
in the math folder there is an init and whoami file, along with two more folders called geometry and numbers. geometry contains it's own init and whoami files along with cube, circle, and rectangle py files. in numbers there is also a init and whoami files along with series and simple py files